// src/app/model/user.ts
export class User {
    id: number;
    username: string;
    _username: string;
    password: string
    email: string;
    national_id: string;
    firstname: string;
    lastname: string;
    phone: string;
    mobile: string;
    birthday: string;
    birthday_sec: string;
    address:string;
    created_at: string;
    created_at_sec: string;
    last_update: string;
    last_update_sec: string;
    level_id: number;
    lang:string;
    block_my_members:string;
    change_price:string;
}